DeviceManager - sleep after recovery actions
Without these, the next call to wifi_timer_callback immediately after a recovery action is sometimes observed to fail the ping because the recovery hasn't quite completed. This is probably benign as we'll just try again on the next wifi_timer_callback, but this makes the logs cleaner by not logging an extra failure.