Don't create /expidite-diags in configuration.py
We have code that runs on startsup to create directories if they don't exist.

Unlike all the other directories, /expidite-diags requires sudo because it isn't on the RAM disk.

We could enhance the code here to use sudo, but this isn't necessary - rpi_installer always creates /expidite-diags so it should always exist. And if it has been deliberately or accidentally deleted, then the next reboot will re-run rpi_installer which will recreate it.