Run `ruff format`
This is a large one off change, but gets the codebase formatting consistent.

Going forwards, we should now run `ruff format` for all commits. This is now documented, along with the other quality checks, in CONTRIBUTING.md (this file is very minimal and should be expanded if we expect external contributions).